Ukończenie implementacji zarządzania certyfikatami serwera OpenVPN
Dokonano następujących zmian:
- Naprawiono inicjalizację mapy Servers w CertificateDB
- Zmieniono typ zwracany GetServerCertificate z error na bool dla konsystencji
- Naprawiono zakres zmiennej daysUntilExpiryInt w handleServerMode
- Stworzono nowy moduł MikrotikIntegration do komunikacji z routerami Mikrotik
- Dodano integrację z Mikrotik API dla automatycznej aktualizacji certyfikatów
- Dobudowano obsługę zmiennych środowiskowych dla danych dostępowych Mikrotika
- Dodano zapisywanie bazy danych w trybie serwera
- Uzupełniono dokumentację i obsługę błędów

Funkcjonalności:
✅ Zarządzanie certyfikatami serwera w Vault
✅ Przechowywanie informacji o certyfikatach w bazie danych
✅ Integracja z API Mikrotik dla automatycznej aktualizacji certyfikatów
✅ Automatyczne wysyłanie certyfikatów na routery
✅ Walidacja i monitoring certyfikatów serwera

+++ internal/mailer.go
@@ -21,7 +21,7 @@ func NewMailer(logger *logrus.Logger) *Mailer {
 	return &Mailer{logger: logger}
 }
 
-func (m *Mailer) SendEmail(fileContent string, expirationDays float64, emailTemplate string, name string, address string) (err error) {
+func (m *Mailer) SendEmail(fileContent string, expirationDays int, emailTemplate string, name string, address string) (err error) {
 	mailer := gomail.NewMessage()
 	mailer.SetHeader("From", os.Getenv("SMTP_FROM"))
 	mailer.SetHeader("To", address)
